update: removed unused code after refactoring
