+++ src/Components/Options/options.types.ts
@@ -1,0 +1,4 @@
+export type OPTION = {
+  option: String;
+  correctOption: String;
+};
