+++ src/Components/QuizQuestion/QuizQuestions.types.ts
@@ -1,4 +1,6 @@
-import { QUESTION } from "../../data/data.types";
+// import { QUESTION } from "../../data/data.types";
+
+import { QUESTION } from "../../types/data.types";
 
 export type QUESTIONS = {
   questions: QUESTION[];
