+++ src/Components/SignIn/SignIn.tsx
@@ -5,7 +5,6 @@ import {
   UserSignIn,
 } from "../../ApiCalls/userAuth";
 import { ServerError, SignInResponse } from "../../types/services.types";
-// import { useQuizPerformance } from "../../context/quizResults.context";
 export function SignIn() {
   const navigate = useNavigate();
   const [userCredentials, setUserCredentials] = useState({
@@ -14,7 +13,6 @@ export function SignIn() {
   });
   const [signinButtonText, setSigninButtonText] = useState("Sign In");
   const [loginError, setLoginError] = useState("invisible");
-  // const { dispatch } = useQuizPerformance();
 
   function SignInSuccess(response: SignInResponse | ServerError) {
     if ("token" in response) {
@@ -22,7 +20,6 @@ export function SignIn() {
       localStorage.setItem("token", response.token);
       localStorage.setItem("username", response.username);
       navigate("/categories", { replace: true });
-      // dispatch({ type: "SET_USERNAME", payload: response.username });
     }
   }
   async function SignInSubmitHandler(event: React.SyntheticEvent) {
