--- src/data/data.tsx
@@ -1,335 +1,0 @@
-// import { CATEGORYDATA } from "../Components/QuizCard/QuizCard";
-
-import { QUIZCATEGORIES } from "./data.types";
-
-export const data: QUIZCATEGORIES = {
-  quizCategories: [
-    {
-      quizName: "Marvel Fanatic",
-      description: "Are you a Marvel fan?If yes then don't miss this",
-      questions: [
-        {
-          question:
-            "IRON MAN 2: What fake name does Natasha use when she first meets Tony?",
-          options: [
-            "Natalie Rushman",
-            "Natalia Romanoff",
-            "Nicole Rohan",
-            "Naya Rabe",
-          ],
-          correctOption: "Natalie Rushman",
-        },
-        {
-          question:
-            "THE AVENGERS: Natasha remarks to Clint that the Battle of New York is a lot like what?",
-          options: [
-            "Their time in Prague",
-            "Their time in Sokovia",
-            "Their time in Budapest",
-            "Their time in Istanbul",
-          ],
-          correctOption: "Their time in Budapest",
-        },
-        {
-          question:
-            "THOR: THE DARK WORLD: Where do Sif and Volstagg hide the Reality Stone at the end of the movie?",
-          options: [
-            "On Vormir",
-            "In a vault on Asgard",
-            "Inside Sif's sword",
-            "They give it to the Collector",
-          ],
-          correctOption: "They give it to the Collector",
-        },
-        {
-          question:
-            "CAPTAIN AMERICA: THE WINTER SOLDIER: What does the Winter Soldier say after Steve recognizes him for the first time?",
-          options: [
-            "Do i know you?",
-            "Who the hell is Bucky?",
-            "He's gone",
-            "What did you say?",
-          ],
-          correctOption: "Who the hell is Bucky?",
-        },
-        {
-          question:
-            "GUARDIANS OF THE GALAXY: What were the three items Rocket claims he needs in order to escape the prison?",
-          options: [
-            "A security card, a fork, and an ankle monitor",
-            "A pair of binoculars, a detonator, and a prosthetic leg",
-            "A Security Band, a battery and a prosthetic leg",
-            "A knife, cable wires, and Peter's mixtape",
-          ],
-          correctOption: "A Security Band, a battery and a prosthetic leg",
-        },
-        {
-          question:
-            "AVENGERS: AGE OF ULTRON: What word does Tony utter that makes Steve say, 'Language'?",
-          options: ["Shit!", "Crap!", "Duffer!", "Idiot!"],
-          correctOption: "Shit!",
-        },
-        {
-          question:
-            "ANT-MAN: What animal does Darren Cross unsuccessfully shrink?",
-          options: ["Mouse", "Sheep", "Duck", "Hamster"],
-          correctOption: "Sheep",
-        },
-        {
-          question: `GUARDIANS OF THE GALAXY VOL. 2: Finish this line: "I'm _______, y'all!" `,
-          options: ["Superman", "Peter Pan", "Underdog", "Marry Poppins"],
-          correctOption: "Marry Poppins",
-        },
-        {
-          question:
-            "BLACK PANTHER: What do Killmonger and Klaue steal from the Museum of Great Britain?",
-          options: [
-            "Just vibranium",
-            "Vibranium and a Mask",
-            "A map of Wakanda",
-            "A map of Wakanda and Vibranium",
-          ],
-          correctOption: "Vibranium and a Mask",
-        },
-        {
-          question:
-            "AVENGERS: INFINITY WAR: What elective class did Thor take on Asgard?",
-          options: [
-            "Learning to speak Groot",
-            "Hammer making",
-            "Learning to fly a space ship",
-            "How to take care of snakes",
-          ],
-          correctOption: "Learning to speak Groot",
-        },
-      ],
-    },
-    {
-      quizName: "Space Wanderer",
-      description: "Hop in if Space fascinates you",
-      questions: [
-        {
-          question: "Who invented the telescope?",
-          options: ["Hypatia", "Johannes Kepler", "Hans Lippershey", "Galileo"],
-          correctOption: "Hans Lippershey",
-        },
-        {
-          question: "Which of these objects is the farthest from the Sun?",
-          options: ["90377 Sedna", "Saturn", "Neptune", "Kuiper belt"],
-          correctOption: "90377 Sedna",
-        },
-        {
-          question: "What is the smallest planet in the solar system by mass?",
-          options: ["Earth", "Mercury", "Mars", "Jupiter"],
-          correctOption: "Mercury",
-        },
-        {
-          question: "What is the visible part of the Sun called?",
-          options: [
-            "the atmosphere",
-            "the photosphere",
-            "the stratosphere",
-            "the lithosphere",
-          ],
-          correctOption: "the photosphere",
-        },
-        {
-          question: "What makes a planet a dwarf planet?",
-          options: [
-            "smell",
-            "color",
-            "size and shape",
-            "distance from the Sun",
-          ],
-          correctOption: "size and shape",
-        },
-        {
-          question: "What two motions do all planets have?",
-          options: [
-            "orbit and spin",
-            "rock and roll",
-            "twist and shout",
-            "wiggle and wobble",
-          ],
-          correctOption: "orbit and spin",
-        },
-        {
-          question: "When did the Space Age begin?",
-          options: ["1957", "1941", "1969", "1999"],
-          correctOption: "1957",
-        },
-        {
-          question:
-            "With which US spacecraft did Neil Armstrong and Edwin Aldrin, Jr. land on the moon?",
-          options: ["Apollo 7", "Apollo 13", "Apollo 17", "Apollo 11"],
-          correctOption: "Apollo 11",
-        },
-        {
-          question: "Who was the first person to enter outer space twice?",
-          options: [
-            "Yuri Gagarin",
-            "Scott Kelly",
-            "Vladimir Komarov",
-            "Christina Koch",
-          ],
-          correctOption: "Vladimir Komarov",
-        },
-        {
-          question: "The atmosphere in Mars is mostly composed of:",
-          options: ["Nitrogen", "Carbon dioxide", "Helium", "Oxygen"],
-          correctOption: "Carbon dioxide",
-        },
-      ],
-    },
-    {
-      quizName: "Science Buff",
-      description: "Dwel into the world of Scientific marvels",
-      questions: [
-        {
-          question:
-            "In 1910 which French engineer and chemist found that passing electrical current through inert gases produced light and proceeded to develop the neon lamp?",
-          options: [
-            "Georges Claude",
-            "Charles Janet",
-            "Edward Adam",
-            "René Jacques Lévy",
-          ],
-          correctOption: "Georges Claude",
-        },
-        {
-          question: "Which of these chemical elements is heavier than iron?",
-          options: ["Gold", "Pottasium", "Carbon", "manganese"],
-          correctOption: "Gold",
-        },
-        {
-          question:
-            "A baby blue whale drinks this many liters of milk per day:",
-          options: ["500", "190", "50", "170"],
-          correctOption: "190",
-        },
-        {
-          question: "Who invented the World Wide Web?",
-          options: [
-            "the Federal Communications Commission",
-            "A laboratory in Switzerland",
-            "Apple",
-            "Microsoft",
-          ],
-          correctOption: "A laboratory in Switzerland",
-        },
-        {
-          question:
-            "Which of the following technological developments came first?",
-          options: ["teletype", "telephone", "telegraph", "telescope"],
-          correctOption: "telescope",
-        },
-        {
-          question:
-            "Who developed the theory of the chemical reactivity of oxygen?",
-          options: [
-            "Antoine Lavoisier",
-            "Louis Pasteur",
-            "John Dalton",
-            "Joseph Priestley",
-          ],
-          correctOption: " Antoine Lavoisier",
-        },
-        {
-          question: "What kind of substance makes litmus paper turn blue?",
-          options: ["Acid", "Salt", "Alkali", "Base"],
-          correctOption: "Alkali",
-        },
-        {
-          question: "Who discovered X-rays?",
-          options: [
-            "Raymond Gosling",
-            "Wilhelm Röntgen",
-            "Riccardo Giacconi",
-            "Marie Curie",
-          ],
-          correctOption: "Wilhelm Röntgen",
-        },
-        {
-          question: "Which of these chemicals help fruit to ripen?",
-          options: ["Ethylene", "Methane", "Nitrogen oxide", "Carbon dioxide"],
-          correctOption: "Ethylene",
-        },
-        {
-          question:
-            "The layer of the atmosphere in which weather occurs is called the:",
-          options: ["Ionosphere", "Troposphere", "Mesosphere", "Stratosphere"],
-          correctOption: " Troposphere",
-        },
-      ],
-    },
-    {
-      quizName: "Aquatic Life",
-      description: "Dive in deep to test yourself aboout what's in there",
-      questions: [
-        {
-          question: "What is the largest fish in the world?",
-          options: ["Whale shark", "Sperm whale", "Sailfish", "Blue whale"],
-          correctOption: "Whale shark",
-        },
-        {
-          question: "What shark can live in both freshwater and saltwater?",
-          options: [
-            "Blue shark",
-            "Blacktip shark",
-            "Great white shark",
-            "Bull shark",
-          ],
-          correctOption: "Bull shark",
-        },
-        {
-          question: "What is the fastest fish in the world?",
-          options: ["Wahoo", "Sailfish", "Mako shark", "Yellowfin tuna"],
-          correctOption: "Sailfish",
-        },
-        {
-          question: "What is the tail fin of a fish called?",
-          options: ["Dorsal fin", "Caudal fin", "Anal fin", "Pectoral fin"],
-          correctOption: "Caudal fin",
-        },
-        {
-          question: "What is a sharks skeleton made of?",
-          options: ["Collagen", "Bone tissue", "Calcite", "Cartilage"],
-          correctOption: "Cartilage",
-        },
-        {
-          question: "Which invertebrate can change genders?",
-          options: ["Oyster", "Blue crab", "Sand dollar", "Barnacle"],
-          correctOption: "Oyster",
-        },
-        {
-          question: "What is a mass of jellyfish floating together called?",
-          options: ["Crash", "Smack", "Unkindness", "Murder"],
-          correctOption: "Smack",
-        },
-        {
-          question:
-            "The exoskeleton of a crab is made primarily of what material?",
-          options: ["Sclerotin", "Keratin", "Chitin", "Concrete"],
-          correctOption: "Chitin",
-        },
-        {
-          question:
-            "What part of a fish may be found in lipsticks, nail polishes and mascara?",
-          options: ["Fin tissue", "Bones", "Scales", "Skin"],
-          correctOption: "Scales",
-        },
-        {
-          question: "How does a sea cucumber defend itself?",
-          options: [
-            "Camouflage",
-            "Hides in crevices",
-            "Throws up",
-            "Secretes ink",
-          ],
-          correctOption: "Throws up",
-        },
-      ],
-    },
-  ],
-};
-// console.log(data);
